renamespace class and extract for SRP
